wip(slice02): S6 verify + docs; move product-storage context hook to facade folder
- Full battery green: product-client test(7)+build, desktop vitest (3099 pass;
  13 failures pre-existing byte-identical on base 06bf880a1), desktop build,
  frontend boundaries, structure --strict (TOTAL 0), git diff --check clean.
- Fix the one strict structure violation introduced by this slice: git-move
  hooks/persistence/use-product-storage-context.ts into a facade responsibility
  folder and rewrite its 12 importers (mechanical; 53 affected tests pass).
- Docs (d1e convention): add web-desktop-client-unification-d1f.md landed
  contract (adapter re-backing/data-preservation decision, grep ledgers,
  deviations, verification matrix); mark d1e complete at #1180 merge and d1f
  current in web-desktop-client-unification.md; add rollout ledger §7 working
  record, advance handoff, renumber external-config gate to §9.
