+++ apps/desktop/src/hooks/persistence/facade/use-product-storage-context.ts
