+++ apps/desktop/src/hooks/preferences/lifecycle/use-user-preferences-lifecycle.ts
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import { useProductStorageContext } from "@/hooks/persistence/use-product-storage-context";
+import { useProductStorageContext } from "@/hooks/persistence/facade/use-product-storage-context";
 import {
   selectPersistedUserPreferencesSlice,
   type PersistedUserPreferencesMetadata,
