+++ specs/codebase/features/web-desktop-client-unification.md
@@ -633,15 +633,27 @@ and clipboard behavior through the mounted host while product source remains in
 Desktop. The complete contract is
 [`web-desktop-client-unification-d1d.md`](web-desktop-client-unification-d1d.md).
 
-Route Shared Identity and Navigation Through ProductHost is the current
-implementation slice. It reads normalized auth identity/operations,
-`host.deployment.apiBaseUrl`, and the single `host.cloud.client` through
-`useProductHost()`, closes the ordered-query-pairs and fragment/callback
-contract gaps in ProductClient, and makes one product-owned lifecycle route
-each inbound deep link decoded once into a lossless `ProductEntry` (the legacy
-parallel navigation decoder is deleted). Product source remains in Desktop. The
-complete living contract is
+Route Shared Identity and Navigation Through ProductHost is complete at PR #1180
+merge `06bf880a1b98c6694bcf029badcc9fe5823111de`. It reads normalized auth
+identity/operations, `host.deployment.apiBaseUrl`, and the single
+`host.cloud.client` through `useProductHost()`, closes the ordered-query-pairs
+and fragment/callback contract gaps in ProductClient, and makes one
+product-owned lifecycle route each inbound deep link decoded once into a
+lossless `ProductEntry` (the legacy parallel navigation decoder is deleted).
+Product source remains in Desktop. The complete contract is
 [`web-desktop-client-unification-d1e.md`](web-desktop-client-unification-d1e.md).
+
+Route Shared Persistence and Telemetry Through ProductHost is the current
+implementation slice. It re-backs `desktopProductStorage` onto the existing
+Tauri preferences store and routes movable product persistence through
+`host.storage` with zero migration of existing values, routes product telemetry
+identity/tags/route-classification/exception-capture and the single
+`screen_viewed` through a product-owned `useProductTelemetry` facade backed by
+`host.telemetry`, makes the Query client product-owned with injected exception
+capture, and splits the mount into a host-owned `DesktopHostProviders`, a
+product `ProductProviderRoot`, and a product `ProductLifecycleRoot`. Product
+source remains in Desktop. The complete living contract is
+[`web-desktop-client-unification-d1f.md`](web-desktop-client-unification-d1f.md).
 
 Related authoritative docs:
